feat(pie-cookie-banner): DSW-000 move default props into defs file
