+++ packages/components/pie-cookie-banner/src/index.ts
@@ -15,8 +15,6 @@ import { PieSwitch } from '@justeattakeaway/pie-switch';
 import { defineCustomElement, dispatchCustomEvent } from '@justeattakeaway/pie-webc-core';
 /* eslint-enable import/no-duplicates */
 
-import defaultLocale from '../locales/en-gb.json';
-
 import styles from './cookie-banner.scss?inline';
 import {
     CookieBannerProps,
@@ -24,29 +22,18 @@ import {
     ON_COOKIE_BANNER_NECESSARY_ONLY,
     ON_COOKIE_BANNER_MANAGE_PREFS,
     ON_COOKIE_BANNER_PREFS_SAVED,
+    defaultProps,
     preferences,
     type Preference,
     type PreferenceIds,
     type CookieBannerLocale,
     type CustomTagEnhancers,
-    DefaultProps,
 } from './defs';
 
 import { localiseText, localiseRichText } from './localisation-utils';
 
 // Valid values available to consumers
 export * from './defs';
-
-// TODO: Move this object to defs.ts
-// It can be moved as soon as we manage to figure a safe way to avoid the error TS2821
-// Import assertions are only supported when the '--module' option is set to 'esnext' or 'nodenext'
-export const defaultProps: DefaultProps = {
-    hasPrimaryActionsOnly: false,
-    defaultPreferences: {},
-    locale: defaultLocale,
-    cookieStatementLink: '',
-    cookieTechnologiesLink: '',
-};
 
 const componentSelector = 'pie-cookie-banner';
